Extracted JsonReader into an interface
