+++ moshi/src/main/java/com/squareup/moshi/JsonReaderInterface.java
@@ -1,0 +1,4 @@
+package com.squareup.moshi;
+
+public interface JsonReaderInterface {
+}
